+++ CHANGES.md
@@ -4,6 +4,7 @@
 - Added some new convenience components, like c-strong c-p c-em which mimic html like structure but with RVO support (thus adding classes)
 - Added the progress tracker component
 - Passthrough for certain attributes like data-, aria-, hx-
+- Fixed custom web data attributes: no conversion from camelcase
 
 # [0.1] - 2025-11-04
 - First tagged version